Begin refactoring into resource and context model
The BuildManifest function has been broken out into components representing the
resources and the context in which a resource is accessed. This keeps resource
resolution out of the manifest code and lets us pull off system specific
aspects. Ideally, the Context becomes an interface, or takes a driver
interface, to facilitate safer testing.

We experiment with an interface based type dispatch models, but we may want to
clean this up. Notes identify problems with the representation.

Device support has been dropped for now, but will come back as we get better
facilities for system specific implementations.

Signed-off-by: Stephen J Day <stephen.day@docker.com>