Refine marshaling and resource creation
A package-specific Manifest type is now exported, requiring protobuf conversion
functions. The package-level interface no longer uses the protobuf types
directly. Implementing the fromProto function led to simplification of the
resource creation functions.

While this is in our desired direction, it's beginning to become apparent that
Resource really does need to operate as a record type. The only incompatibility
with proto.Resource is the type of each field. We may consider updating the
types and field names of the generated protobuf to better match Go's types,
rather than have multiple types that are doing the same thing. Before going in
this direction, we must be sure that coupling these types will not lead to
future problems.

Signed-off-by: Stephen J Day <stephen.day@docker.com>