apply: correct handling of symlinks
Previously, tricky things were attempted with symlinks to make filesystems more
portable. Without operating system support, this effort was wasted and caused
bugs. Code attempting to fixup symlinks has been elided in favor of simply
transporting the link targets as is.

A small bug in lchown was fixed, as well.

Signed-off-by: Stephen J Day <stephen.day@docker.com>