Provide a constant for ENODATA
There is no real standardisation between Unix versions as to whether
ENODATA or ENOATTR is used, so provide a standard constant in sysx that
can be used by consumers.

Linux returns ENODATA, while FreeBSD and Darwin return ENOATTR; to
confuse things ENODATA is defined on Darwin but is not returned for
extended attributes.

Applications should use `sysx.ENODATA` instead.

Signed-off-by: Justin Cormack <justin.cormack@docker.com>