Fix copy_file_range usage for files > 2GB on 32-bit archs.
copy_file_range checks that its arguments don't exceeds SSIZE_MAX and fails with EINVAL if they do.

This fix ensures that we are never trying to copy more than SSIZE_MAX at a time and at the same time avoids overflows when the file is larger than 4GB on 32-bit systems.

Signed-off-by: Kasper Fabæch Brandt <poizan@poizan.dk>