freebsd: use utimensat syscall in favor of utimes
The utimes syscall follows symlinks and changes the times on the link
target, rather than the link itself.  The lutimes syscall can be used to
modify times on a symlink, though both syscalls are deprecated in favor
of utimensat.  utimensat can be called with the AF_SYMLINK_NOFOLLOW flag
to change the times on a symlink.

Signed-off-by: Samuel Karp <me@samuelkarp.com>