Allow Walk to visit volume mounts on Windows
This change adds a fork of the filepath.Walk function from the standard
library, with changes that allows it to visit reparse points that are
in fact volume mounts.

Currently go makes no distinction between a symlink, directory junction,
volume mount or Unix domain socket. All reparse points are regarded as
symlinks. This means that FileMode does not have ModeDir set for any
path that also has ModeSymlink. This includes volume mounts. See:
https://go-review.googlesource.com/c/go/+/41830/

This implementation also attempts to avoid recursions.

Fixes: #174

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>