Add paths to windows metadataFiles
When creating a new empty base layer on Windows, hcsshim creates a set
of empty registry hives in C:\Windows\System32\config. These files are
needed in order for the new layer to be considered valid.

Add these locations to the metadataFiles, which excludes them when
comparing the diffs of two layers.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>